Restructure the crate
