+++ src/algorithms/optimal_string_alignment.rs
@@ -1,7 +1,7 @@
 use std::cmp::min;
 use std::mem;
 
-use crate::fuzzy::interface::{Similarity, SimilarityMetric};
+use crate::algorithms::{Similarity, SimilarityMetric};
 
 /// Like Levenshtein but allows for adjacent transpositions. Each substring can
 /// only be edited once.
